added back helix and particle metadata I accidentally removed

+++ programs/e2helixboxer.py
@@ -60,7 +60,18 @@ def get_particles_from_helix( helix, px_overlap, px_length = None, px_width = No
     (x0,y0) = (0,0)
     particles = []
     while y0 + ysize <= helix.get_ysize():
-        particles.append( helix.get_clip(Region(x0, y0, xsize, ysize)) )
+        ptcl = helix.get_clip(Region(x0, y0, xsize, ysize))
+        if helix.has_attr("helix_source_coords"):
+            hcoords = helix.get_attr("helix_source_coords")
+            l_vect = (hcoords[2]-hcoords[0], hcoords[3] - hcoords[1])
+            length = sqrt(l_vect[0]**2+l_vect[1]**2)
+            l_uvect = (l_vect[0]/length, l_vect[1]/length)
+            x1 = hcoords[0] + y0*l_uvect[0]
+            y1 = hcoords[1] + y0*l_uvect[1]
+            x2 = x1 + px_length*l_uvect[0]
+            y2 = y1 + px_length*l_uvect[1]
+            ptcl["helix_ptcl_source_coords"] = (x1, y1, x2, y2, px_width) 
+        particles.append( ptcl )
         y0 += px_step
     
     return particles
@@ -94,6 +105,7 @@ def get_helix_from_coords(micrograph, x1, y1, x2, y2, width):
     tr.set_rotation({"type":"2d", "alpha":rot_angle})
     helix_dimensions = ( int(round(width)), int(round(length)), 1 )
     helix = micrograph.get_rotated_clip( tr, helix_dimensions )
+    helix["helix_source_coords"] = (x1, y1, x2, y2, width)
     return helix
 
 def load_coords(coords_filepath):
@@ -178,6 +190,7 @@ def db_get_helices_dict(micrograph_filepath):
     helices_dict = {}
     for coords in box_coords_list:
         helix = get_helix_from_coords(micrograph, *coords)
+        helix["ptcl_source_image"] = micrograph_filepath
         helices_dict[tuple(coords)] = helix
     return helices_dict
 
@@ -204,7 +217,7 @@ def db_save_helices(micrograph_filepath, output_type = "hdf", output_dir=""):
         save_helix(helix, micrograph_name, i, output_type, output_dir)
         i+=1
 
-def save_particles(helix_emdata, helix_num, ptcl_filepath, px_overlap, px_length = None, px_width = None, do_edge_norm=False):
+def save_particles(helix_emdata, helix_num, ptcl_filepath, px_overlap, px_length = None, px_width = None, do_edge_norm=False, as_stack=True):
     """
     saves the particles in a helix to a stack file
     @helix_emdata: the EMData object containing the image data for the helix
@@ -214,14 +227,15 @@ def save_particles(helix_emdata, helix_num, ptcl_filepath, px_overlap, px_length
     """
     particles = get_particles_from_helix(helix_emdata, px_overlap, px_length, px_width)
     (path, ext) = os.path.splitext(ptcl_filepath)
-    ptcl_filepath = "%s_%i%s" % (path, helix_num, ext)
-    if os.access(ptcl_filepath, os.F_OK):
-        os.remove(ptcl_filepath)
-    for i in range(len(particles)):
-        ptcl = particles[i]
-        if do_edge_norm:
-            ptcl.process_inplace("normalize.edgemean")
-        ptcl.write_image(ptcl_filepath, i) #appending to the image stack
+    if as_stack:
+        ptcl_filepath = "%s_%i%s" % (path, helix_num, ext)
+        if os.access(ptcl_filepath, os.F_OK):
+            os.remove(ptcl_filepath)
+        for i in range(len(particles)):
+            ptcl = particles[i]
+            if do_edge_norm:
+                ptcl.process_inplace("normalize.edgemean")
+            ptcl.write_image(ptcl_filepath, i) #appending to the image stack
 
 
 def db_save_particles(micrograph_filepath, px_overlap, px_length = None, px_width = None, output_type = "hdf", output_dir=""):
@@ -272,6 +286,10 @@ class EMWriteHelixFiles(QtGui.QDialog):
         self.ptcls_groupbox.setCheckable(True)
         self.ptcls_edgenorm_checkbox = QtGui.QCheckBox(self.tr("&Normalize Edge-Mean"))
         self.ptcls_edgenorm_checkbox.setChecked(False)
+        self.ptcls_filename_line_edit = QtGui.QLineEdit()
+        self.ptcls_as_stack_checkbox = QtGui.QCheckBox(self.tr("&As Stack"))
+        self.ptcls_as_stack_checkbox.setChecked(True)
+        self.ptcls_as_stack_checkbox.setEnabled(False)
         ptcls_overlap_label = QtGui.QLabel(self.tr("&Overlap:"))
         self.ptcls_overlap_spinbox = QtGui.QSpinBox()
         self.ptcls_overlap_spinbox.setMaximum(1000)
@@ -288,7 +306,6 @@ class EMWriteHelixFiles(QtGui.QDialog):
         self.ptcls_dir_line_edit = QtGui.QLineEdit()
         self.ptcls_browse_button = QtGui.QPushButton(self.tr("Browse"))
         ptcls_filename_label = QtGui.QLabel(self.tr("File Name:"))
-        self.ptcls_filename_line_edit = QtGui.QLineEdit()
         
         self.button_box = QtGui.QDialogButtonBox(QtGui.QDialogButtonBox.Save | QtGui.QDialogButtonBox.Cancel)
         
@@ -331,6 +348,7 @@ class EMWriteHelixFiles(QtGui.QDialog):
         
         ptcls_opts_layout = QtGui.QVBoxLayout()
         ptcls_opts_layout.addWidget(self.ptcls_edgenorm_checkbox)
+        ptcls_opts_layout.addWidget(self.ptcls_as_stack_checkbox)
         ptcls_opts_layout.addLayout(ptcls_overlap_layout)
         ptcls_opts_layout.addLayout(ptcls_width_layout)
         ptcls_opts_layout.addLayout(ptcls_length_layout)
@@ -579,6 +597,7 @@ class EMHelixBoxerWidget(QtGui.QWidget):
             key = self.generate_emshape_key()
             self.main_image.add_shape(key, emshape)
             helix = get_helix_from_coords(self.main_image.get_data(), *coords)
+            helix["ptcl_source_image"] = self.micrograph_filepath
             self.helices_dict[coords] = helix
             self.add_box_to_db(coords)
 
@@ -715,7 +734,8 @@ class EMHelixBoxerWidget(QtGui.QWidget):
             old_coords = old_emshape.getShape()[4:9]
             new_coords = (old_coords[0], old_coords[1], old_coords[2], old_coords[3], width)
             helix = get_helix_from_coords( self.main_image.get_data(), *new_coords )
-            
+            helix["ptcl_source_image"] = self.micrograph_filepath
+                        
             self.remove_box_from_db(old_coords)
             self.add_box_to_db(new_coords)
             self.helices_dict.pop(tuple(old_coords))
@@ -900,6 +920,7 @@ class EMHelixBoxerWidget(QtGui.QWidget):
                     self.edit_mode = "2nd_point"
                     
                     helix = get_helix_from_coords( self.main_image.get_data(), *self.initial_helix_box_data_tuple )
+                    helix["ptcl_source_image"] = self.micrograph_filepath
                     self.display_helix(helix)
                 
             elif self.edit_mode == "delete":
@@ -928,6 +949,7 @@ class EMHelixBoxerWidget(QtGui.QWidget):
                 
                 box_coords = tuple( box.getShape()[4:9] )
                 helix = get_helix_from_coords( self.main_image.get_data(), *box_coords )
+                helix["ptcl_source_image"] = self.micrograph_filepath
                 self.display_helix(helix)
 
     def mouse_up(self, event, cursor_loc):
@@ -945,6 +967,7 @@ class EMHelixBoxerWidget(QtGui.QWidget):
             box = self.main_image.get_shapes().get(self.current_boxkey)
             box_coords = tuple( box.getShape()[4:9] )
             helix = get_helix_from_coords( self.main_image.get_data(), *box_coords )
+            helix["ptcl_source_image"] = self.micrograph_filepath
             self.helices_dict[box_coords] = helix
             
             self.add_box_to_db(box_coords)
